Versión 1.10.6:     - ImagenCommons: imagenen predeterminada ahora se carga mediante un drawable.
