+++ app/src/main/java/es/icp/icp_commons/Objects/ImagenCommons.java
@@ -1,7 +1,10 @@
 package es.icp.icp_commons.Objects;
 
+import android.content.Context;
 import android.graphics.Bitmap;
 
+import es.icp.icp_commons.R;
+import es.icp.icp_commons.Utils.CommonsUtilsImagenes;
 import es.icp.icp_commons.Utils.Localizacion;
 import es.icp.icp_commons.Utils.UtilsFechas;
 
@@ -12,8 +15,8 @@ public class ImagenCommons {
     private int                   orientacion;
     private String                fecha;
 
-    public ImagenCommons() {
-        this("/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAoHBwkHBgoJCAkLCwoMDxkQDw4ODx4WFxIZJCAmJSMgIyIoLTkwKCo2KyIjMkQyNjs9QEBAJjBGS0U+Sjk/QD3/2wBDAQsLCw8NDx0QEB09KSMpPT09PT09PT09PT09PT09PT09PT09PT09PT09PT09PT09PT09PT09PT09PT09PT09PT3/wgARCAIABAADAREAAhEBAxEB/8QAGwABAAIDAQEAAAAAAAAAAAAAAAYHAQMFBAL/xAAaAQEAAwEBAQAAAAAAAAAAAAAAAQQFAwIG/9oADAMBAAIQAxAAAAC5QAYRlIwZAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAB8ojdmpHLVTxdOZAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAGT28+klq3ZFWt5AAAAAAAAAAAAAAAAAAAAAAAAAAAACPH78VfqY3EtVgQAAAAAAAAAAAAAAPdx7S6neh1+h5PfIAAAAAAAAAAAAAAkgmQVbdm5Wx6vPsAAAAAAAAAAAAAAAAAAAAAAAAAAAap81Hs4fJ71iSAAAAAAAAAAAAAAAiZ7nak5z9KvtLKhWhmgAAAAAAAAAAAAAAAd2tbtjH2/pIAAAAAAAAAAAAAAAAAAAAAAAAAAEKu51e6mUQAAAAAAAAAAAAAAAET7PPuVUb8SvUPN75gAAAAAAAAAAAAAAAE2Xk7ErqXcpAAAAAAAAAAAAAAAAAAAAAAAAAAFRbOFxbFYgIkAAAAAAAAAAAAAAAkgAAAAAAAAAAAAAAABMASKrbtTI3MgAAAAAAAAAAAAAAAAAAAAAAAAABFI7vz3m6cgTaePtdjhYSAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHAsVaz1scEdLj3uXE3yQAAAAAAAAAAAAAAAAAAAAAAAAACKP3vnNHvyQTbeLu9zhZQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEds1Kq2cMDoce9z4m+SAAAAAAAAAAAAAAAAAAAAAAAAAARR+987o9+CCbcxd3t8LOQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAARyzUqrYxCB0OPa58T6AkAAAAAAAAAAAAAAAAAAAAAAAAAAij9753R78EE25i7vb4WcgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAjlmpVWxiEDoce1z4n0BIAAAAAAAAAAAAAAAAAAAAAAAAABFH73zuj34IJtzF3e3ws5AAAAAAAAAAAAAAAAAAAAAAAAAAAAAABHLNSqtjEIHQ49rnxPoCQAAAAAAAAAAAAAAAAAAAAAAAAACKP3vndHvwQTbmLu9vhZyAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACOWalVbGIQOhx7XPifQEgAAAAAAAAAAAAAAAAAAAAAAAAAEUfvfO6Pfggm3MXd7fCzkAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEcs1Kq2MQgdDj2ufE+gJAAAAAAAAAAAAAAAAAAAAAAAAAAIo/e+d0e/BBNuYu72+FnIAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAI5ZqVVsYhA6HHtc+J9ASAAAAAAAAAAAAAAAAAAAAAAAAAARR+987o9+CCbcxd3t8LOQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAARyzUqrYxCB0OPa58T6AkAAAAAAAAAAAAAAAAAAAAAAAAAAij9753R78EE25i7vb4WcgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAjlmpVWxiEDoce1z4n0BIAAAAAAAAAAAAAAAAAAAAAAAAABFH73zuj34IJtzF3e3ws5AAAAAAAAAAAAAAAAAAABgQSQCQGQAAAAAARyzUqrYxCB0OPa58T6AkAAAAAAAAAAAAAAAAAAAAAAAAAAij9753R78EE25i7vb4WcgAAAAAAAAAAAAAAAAAAAwef14jFqnxe9fz+/Hp8e+3wsSarc9Hn2AAAAAAI5ZqVVsYhA6HHtc+J9ASAAAAAAAAAAAAAAAAAAAAAAAAAARR+987o9+CCbcxd3t8LOQAAAYEQmQBkAGIJAIJZMBCJzIAYEEhkGIYmIhcowDSy/P75gAenx0sHM1ZbTu5mQAAAABHLNSqtjEIHQ49rnxPoCQAAAAAAAAAAAAAAAAAAAAAAAAACKP3vndHvwQTbmLu9vhZyAAADi9q/P6ckh6+fTsce2+PYA4fet4ffPEgh6PPqRV7eufMZs1Pb4997haAGEeX15j9ir0ufXs8bAwiAaGZB9DNTAABKCYnubqzqhpZAAAAAI5ZqVVsYhA6HHtc+J9ASAAAAAAAAAAAAAAAAAAAAAAAAAARR+987o9+CCbcxd3t8LOQAAAVxpZHAs1uvx7/MxzunLwe+c8z9OZ0r/ANJFZamNxu/Dp8uwQ6HPtYWbp+Tp4pXc+d3+el1Yf0GyPWQYINfzYJo5cxo6NiZuqIzZp1dr42J8j3cuuJeL3zISAzE2tj7kgrWsyAAAAEcs1Kq2MQgdDj2ufE+gJAAAAAAAAAAAAAAAAAAAAAAAAAAIo/e+d0e/BBNuYu72+FnIAAAK40sj7j1YedqgjhWK9XauLP8AO1JhTvisdTG6HPtPc/TGEDMPJ08U3tYHv8dJvnaUprXcg+UU1tYPu8e+ry72Jm6vyintrB5XeuB3q1q0MnarvSyorcopEAdmvYt7G38gAAAAjlmpVWxiEDoce1z4n0BIAAAAAAAAAAAAAAAAAAAAAAAAABFH73zuj34IJtzF3e3ws5AAABXGlkbHqw83VAENuZ8H0M658T6D6iay1MboeOs9ztQBAeTpzpfb+fn2fpx6zUtbI3Mg4FirXulldqvY+3qw8zV5fbhTm3gJgDvVrVs4+6RF7VOutTJ8vvwQETdGF9D0PHUAAAARyzUqrYxCB0OPa58T6AkAAAAAAAAAAAAAAAAAAAAAAAAAAij9753R78EE25i7vb4WcgAAArjSyNj1YebqgDw++VK7vz9z4W/0PHWstTI2wmNLQA9Pjp7ufTydOdL7Xz90Yv0FMbWBceNu9Dn2FZamN2eFnl9eOxNh5urFrdKsdbFAHerWrZx90Dx++dc6eTGrVNMC08fckta2AAAAI5ZqVVsYhA6HHvc2Jv5SAAAAAAAAAAAAAAAAAAAAAAAAAARR+987o9+CCbcxd3t8LOQAAAVxpZGx6sPN1QBo9eaP3fnLgxd7rce9Za2PwLFT1+OgIllLQmtHQ8nTnS+3gXlg/RVjq4/V4951Q0vP68UztYFy4u9AdHM2efVh5urE7tGtNXGAHerWrZx93JgI5HbhVOvieXp4RFoZG5KK1sYhiY5XXjqnyABxe9evdTKIHt5dbPydkAAZMAJTCAAAAAAAAAJ6fPrv8+8gAAAAAAAAAAIo/e+d0e/BBNuYu72+FnIAAAK40sjY9WHm6oA8HvnS2787dOF9D7vHWstTG6HjtPc7TBBKHk6c6X28C8sH6KPWK1damRcuJ9BELdHhWKtnZWzXGpkbPPqw83V4VitUu1hEAd6tatnH3R8ohF/Ngmhm/HryiSLfxfoOzx7jCK208mI3aCYJBABJARKYAAAJyTbP0Y3Zq8zvXAAAAAAABII9XLtbeNudLx0JAAAAAAAAAAIo/e+d0e/BBNuYu72+FnIAAAK40sjY9WHm6owCGXc+FX8+58Pf+k1jqY3Q59p9n6YAxDydOdL7eBeWD9F8opncwLKy9ev9HLn+bpyDhbrjTyNkerDzdXX680pufPef34IHerWrZx93mdeNa6eRxbNYgD18+t04f0GyPQ1z5ozf+dxPgAAAAAAAADb59XPh/QwO/mQ+/ngAAAAAAAACwcvWm1LRAAAAAAAAAABFH73zuj34IJtzF3e3ws5AAABXGlkbHqws3VyYRwLFartbGsHN1ZhTvCsdTG6HPtPs/TAGIeTpzpfbwLywfospgOhlx+1U80rlxd76hXGnk7I9WHm6ogV/Mgullgd2tZk1S9AdHK1e4IAJn+ZqTijpZBqnzRn0Hzh5H359ZSiEgAAAAACRnzOPUIhIAAAAAAIE/HvzhAn+ZrTejo5AAAAAAAAAACKP3vndHvwQTbmLu9vhZyAAACuNPI4fet2+Fn4Ryu3HndOU7z9KaUtD6SKx1Mfl96/X4WBiJ9/jrPM/R8nvxS+38/eOD9H9J5vTjTe789O83UnlDTFb6eRsj1YmbqjVPmpdjD43esllOAgAInuV7VsZG19x6yDVPmjN/wCcT5Fq4+3Ia1vIAAAAAAAAAAAAAAAABTW3gc3twE/zdWcUNIAAAAAAAAAAEUfvfO6Pfggm3MXd7fCzkAAAHB71uZ24jMPb46dvhZ3+fQyCPWKnP6c0gR6fHuT1bmqYitqnLql3KSIrap9+va9nn2OD3rZO7wsjCPN681jq48ftVAQACZHUt2Zl7Ho8+xkGqfNGb/zifItXH3JFWt5AAAAAAAAAAAAAAAABhFNbeBze3AT/ADdWcUNIAAAAAAAAAAEUfvfO6Pfggm3MXd7fCzkAAAAAAAAAAAAAAAAAAAHwiMWqcRu0ONYr6kbY9drh3l9LQk1W39JAA1T5ozf+cT5Fq4+5Iq1vIAAAAAAAAAAAAAAAAMIprbwOb24Cf5mrOKOkAAAAAAAAAACKP3vndHvwQTbmLu9vhZyAAAAAAAAAAAAAAAAAAADACMHxMfcTlKGZAADVPmjN/wCcT5Fq4+5Iq1vIAAAAAAAAAAAAAAAAMIprbwOb24Cf5mrOKOkAAAAAAAAAACKP3vndHvwQTbmLu9vhZyAAAAAAAAAAAAAAAAAAAAAAAAAADVPmjN/5xPkWrj7kirW8gAAAAAAAAAAAAAAAAwimtvA5vbgJ/mas4o6QAAAAAAAAAAIo/e+d0e/BBNuYu72+FnIAAAAAAAAAAAAAAAAAAAAAAAAAANU+aM3/AJxPkWrj7kirW8gAAAAAAAAAAAAAAAAwimtvA5vbgJ/mas4o6QAAAAAAAAAAIo/e+d0e/BBNuYu72+FnIAAAAAAAAAAAAAAAAAAAAAAAAAANU+aM3/nE+RauPuSKtbyAAAAAAAAAADAAAAMgAwimtvA5vbgJ/mas4o6QAAAAAAAAAAIo/e+d0e/BBNuYu72+FnIAAAAAAAAAAAAAAAAAAAAAAAAAANU+aM3/AJxPkWrj7kirW8gAAAAAAAAAAwiPWKsOu0Od24+jx7k9O7Mal77j1kAGEU1t4HN7cBP8zVnFHSAAAAAAAAAABFH73zuj34IJtzF3e3ws5AAAAAAAAAAAAAAAAAAAAAAAAAABqnzRm/8AOJ8i1cfckVa3kHO98uL3rpe3n07fCx9J+URe1UxMDrcO/s8+4/3rSetb+0jy+vEf71pTWtkwa9mwHSzE+QB3K1u1sja3R6AGEU1t4HN7cBP8zVnFHSAAAAAAAAAABFH73zuj34IJtzF3e3ws5AAAAAAAAAAAAAAAAAAAAAAAAAABqnzRm/8AOJ8i1cfckVa3kESt0eT34d6tZ5PbhxO/C0MrYyVJsYc6z9IdPl20+vNU6+HaOTt97hZEIvZ0MvZ90Yf0HG7V6j2sLExJKd3y9OfFsVyEJxQ07AztQAYRTW3gc3twE/zNWcUdIAAAAAAAAAAEUfvfO6Pfggm3MXd7fCzkAAAAAAAAAAAAAAAAAAAAAAAAAAGqfNGb/wA4nyLVx9yRVreQRK3REtqXsENu5x7l1O5V2rjWrk7QJ4/avDrtAWTma+EVhqZGj14tbJ2q60sqHX84TTP0rAztSEX82C6Ob8I9fjpdeF9D9JAwimtvA5vbgJDVud2vZAAAAGSQVrfY49wAAAARR+987o9+CCbcxd3t8LOQAAAAAAAAAAAAAAAAAAAAAAAAAAap80Zv/OJ8i1cfckVa3kESt0RLal4cfrXh96jYebqUztYHd4WhNqGjifMds1OP242Zma3K68eH3rx+zVtbI2qr18WO2qhE0z9Gw87WwR+xUqvYxMJvHB+i2+feQYRTW3gc3twCCZAAAAA2+Pd1Ye/6PPQAAAAij9753R78EE25i7vb4WcgAAAAAAAAAAAAAAAAAAAAAAAAAA1T5ozf+cT5Fq4+5Iq1vIIlboiW1LuEx+xUj9mvPM/Rq7UxrVytoE8jtXjlmp7+fb4mON2rzvP0a01cm1sjarnTyYfezxNKGjYebrRuzTrfUyfJ05+rn7uzC+i+kgYRTW3gc3twAARMzoaHs8dcnxPmBaOYmAETcuH9B0ufYAAAAij9753R78EE25i73b4WMgAAAAAAAAAAAAAAAAAAAAAAAAAA1T4ozf8AnU+RauPuSKtbyCJW6IllS9o9eKw1MefZ2n1OfartTHtbK2hg4/avHbVSa0NGrNbF9nPtZmXq1RsY1rZGzw+9eptnCxMS2lfzExG9QBE0z9OxM7UAGEU1t4HN7cAAB3K1nd595PmYj9mriYJIRNy4f0HS59gAAACKP3vndHvwQO3Wtejz6JIAAAAAAAAAAAAAAAAAAAAAAAAAAxKO2qYFq4+5Iq1vIIvZpxC9n+nx71yl9K9Iq1vX681Hr4nS59UpTTu9Dn04NitM6WhX+hl96vZkVa3WWrkWbl6xNf6OVCNDOCYJQ69exbGRt+nz7AGEU1t4HN7cUROM/T7nCyAAAAPN7515o5du4270ufYAAAAij9753R78EBBIAAAAAAAAAAAAAAAAAAAAAAAAAAAlEJC1cfckVa3kGIJEEgAEBEkJEkEgjKRiEXt04dez+b04+nx7k1W5N6OjticpAGEU1t4HN7cfR493dg/RE5AAAABWunjzihpdLn2AAAAIo/e+d0e/BAAAAAAAAAAAAAAAAAAAAAAAAAAAAAJIAtXH3JFWt5AAAAAAAAAAAAMAyAAAYRTW3gc3tw9fPrdmF9DlIAAAAFbaWRNqWj0uXYAAAAij9753R78EAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAC1cfckVa3kAAAAAAAAAAAAAAAAGEU1t4HN7cPXz63ZhfQ5SAAAABW2lkTalo9Ll2AAAAIpLd+e8vXmiAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABauPuSKtbyAAAAAAAAAAAAAAAADCKa28Dm9uHr59bswvocpAwBBIZAK20sibUtHpcuwAAAAqfXw+BaqkAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAC1cfckVa3kAAAAAAAAAAAAAAAAGEU1t4HN7cPXz63ZhfQ5SMEWtU4jdz9Prx3q1ucUdHf595BW2lkTalo9Ll2AAAAETtUq118XAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAALVx9yRVreQAAAAAAAAAAAAAAAAYRTW3gc3tw9fPrdmF9DlOCFXc6vtPK3efWj34Q7Na1bWTt7Y9CttLIm1LR6XLsAAAAPlFWa2NHbdMgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAC1cfckVa3kAAAAAAAAAAAAAAAAGEU1t4HN7cPXz63ZhfQ5TpnzSm789o987Lydnm9eUI0M35nzZ2PuSqtbFbaWRNqWj0uXYAAAADVPmutLKid2jifKJSAAAAAAAAAAAAAAAAAAAAAAAAAABCJTAtTG25FXuZQAAAAAAAAAAAAAAAAMIpvbweb24evn1uzC+hynm++NM73z5FnZG1KatyL2qdY62NOc/Tn2fpittLIm1LR6XLsAAAAAMHM68Y7ZqeLpyQAAAAAAAAAAAAAAAAAAAAAAAAAAHzMQ29QTA9Hjp9xKYAAAAAAAAAAAAAAAAQJ83rn8+o9fPrdmF9DlOifFJ7vz+r34s7I2cSrrTyfP652XlbMtqXhW2lkTalo9Ll2AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAGufFGb/wA6QAAAAAAAAAAAAAAAAAAAEx6+fW7ML6HKcEDv5kE0svpce/N68Ux1eFi3cbd3R6FbaWRNqWj0uXYAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAa58Uju/PkJAAAAAAAAAAAAAAAAAABAbo9XZhfQ5T8oEPu0Ijco6fXnvVrM8z9P1+emQVtpZE2paPS5dgAAAAAAAAAAAAAAAAPmYrbSyfmYIAAAAAAAAG6PdjZmrlIAAAAABAAAAAAAAAAAAAAAAAAAAGCptfEtrI28pAGAZAAK20sibUtHpcuwAAAAAAAAAAAAAAAAwczpxpzc+fSIAAAAAAAADz6uvD+h9njoABiYwgAAAAAAAAAAAAAAAAAAAAAAACp9bFtrI28pAAAAArbSyJtS0ely7AAAAAAAAAAAAAAAADBE7dGtdbGAAAAAAAAACFp5O3JKtwAYKw1ceNWqYAAAAAAAAAAAAAAAAAAAAAAAI9Hn3duF9DlIAAAAFbaWTNqOh0ufYAAAAAAAAAAAAAAAAYhXWnkw6/nAAAAAAAAAATrO057n6gAwQW/mwPSyyAAAAAAAAAAAAAAAAAAAAAAASR6PHS7sH6L6SAAAABWenjzmhpdLn2AAAAAAAAAAAAAAAHM6coZeztXrxH7NXx9OYAAAAAAAAARPv8dO7Ws7/AD0mtDQ93jrk+J8wW9nRG9Q8fvwQAAAAAAAAAAAAAAAAAAAAAABMKN/t8LIAAzEgDzdOcCv5tv4+70uXYAAAAAAAAAAAAAAAU9tYHJ78ACAAAAAAAAAAESkR3atq2sjdJIHzEeb348/rykAAAAAAAAAAAAAAAAAAAAAMIqvWxvB24AkESene63DvA9PLwgBEpi5cL6Hpc+wAAAAAAAAAAAAAABFIb3zvn98wSQAAAAAAAAAAAOjx7XLifQ5AAAAAAAAAAAAAAAAAAAAAAAABhFN7eDze3AEAWti7vd4WKU3vnvP75kghE3LifQdLl2AAAAAAAAAAAAAAAFV62LHLVNMAAAAAAAAAAAACV0b9l5exmQAAAAAAAAAAAAAAAAAAAAAAAAwVfq40ZuUQAifV59+nn75nfgACfRz6XRib+/z7AAAAAAAAAAAAAAAHi9847Zq4QEgAAAAAAAAAAhkktW56fPvIAAAAAAAAAAAAAAAAAAAAAAAABo9eI1Zq6/XkAAABAd2vZ6fPsAAAAAAAAAAAAAAABgyAAAAAAAAAAAAAYRlIAAAAAAAAAAAAAAAAAAAAAAAABBIAAAAABBIAAAH/xABNEAAABAIBDQwIBAUFAQADAAABAgMEAAUGBxESFhc2U1Ryc5GSkxATFSAhMTVRUlWxshQwMzRBUHHRIjJAdEJgYaHCI2OBgsFwYqPh/9oACAEBAAE/AP8A4IYwEKImEAAPiMTOncsYVyIV3agYPm0w5qjTFX3dBBHSaLep5jKeyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvU8xkmyLFvc8xhPZBDWqM/J702QWyK5YlVN5XMayahxbLdlX7wAgIVw+dzOZt5SzO5dHsSBpMPUET6lLydqGLXFFr8Eij4/IZPKHM6eg3a/U5x5ihFzUm8dIG37N8kTKWuJU9O1dFrHLoMHWHyGj1LncmOCSoiuz7A85cmGL9CYtCOWqgHSPzD85WWI3QOsqYCpkATGEfgEUinys9mIq8yBORInyKpqKe8v8AC2RdXcqlCl6WywtgbR8iotSE8jf1jmrtFeRUv+UEOVQgHIICUwVwEPj84qiTYUGiUuS51/xqZIfI5ZNHMpeA5aHrH5hD4GDqGLpS28cjAgLZzkh+/cTJ4dy6PZqn/t/T5HQCbC8lJ2ao11GvkH5xS52Lykzw3wTMCRf+OJXCK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xXDriuHXFcOuK4dcVw64rh1xX4lBXYtaTJE+C5TJj4h84mJhPNHZ+tY4/3HiUSlzNajDI6rRA5xINcxkw6xjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFjgmX4i22RY4Jl+IttkWOCZfiLbZFimrBohRlc6LVBM9kTlKmHa4lHTCWkcvz5fnD7pBznTePEobeoxyB8w/yBTu9RxlE83EkF8Mv/cE+cPukHOdN48Sht6jHIHzD/IFO71HOWTzcSQXwy/9wT5w+6Qc503jxKG3qMcgfMP8gU7vUc5ZPNxJBfDL/wBwT5w+6Qc503jxKG3qMcgfMP8AIFO71HOWTzcSQXwy/wDcE+cPukHOdN48Sht6jHIHzD/IFO71HOWTzcSQXwy/9wT5w+6Qc503jxKG3qMcgfMP8gU7vUc5ZPNxJBfDL/3BPnD7pBznTePEobeoxyB8w/yBTu9Rzlk83EkF8Mv/AHBPnD7pBznTePEobeoxyB8w/wAgU7vUc5ZPNxJBfDL/ANwT5w+6Qc503jxKG3qMcgfMP8gU7vUc5ZPNxJBfDL/3BPnD7pBznTePEobeoxyB8w/yBTu9Rzlk83EkF8Mv/cE+cPukHOdN48Sht6jHIHzD8rdvm7FIVHKxEidZhh/VDYIcjRJRcdAQvVGmB/YtkU4t/nPWhqQhVGfk9u2QUhhVDl6/I6IdsbSENniD1IFGypFSD8Sj6+nd6jnLJ5uJIL4Zf+4J84fdIOc6bx4lDb1GOQPmH5VSKnKTETNpbWWX+J/gSHj9zMFhVdrHVMPXxmMxdS1YFWax0zf2GKOU3RmVZs/rIufgb4H9dTu9Rzlk83EkF8Mv/cE+cPukHOdN48Sht6jHIHzD6qc0nZSNciLvfLI5bILEsXQpR/v6kXQ5R/v6kXQpOOH2cN6aSVfmdgTLhB0g5JZoKkUL1lGvxZvSthJXYN3W+WYlr/hLF0KUf7+pF0KUf7+pF0KUf7+pF0KUf7+pDCmstmL1Nqhv1mp1k3F1it0Dqn/KQomGLoUoHD6kSql8um70GrcVAUHtF40wfoy1ko6cDWTTCuMXQ5R1L6kSWkrOeqqkaWddMK42ReJTKlg2Z5dLj51QPAPVUNpaKoll0xPmlB8B9bTu9Rzlk83EkF8Mv/cE+cPukHOdN48Sht6jHIHzD6qqR0q0zQxJpE6niqpGgpgKYVxs4uezftoQagE4DAa8PKLTdkAiqyOYvWT8UNXrqXrWbZZRE4dQ1oo3Tkrw5Wk0rEV5iq/A0AO7VDvgTzMSeQPJ3vvoYEHeq1lZGi0Cc9SGvFoE56kNeLQJz1Ia8SGhs0l87bOlwRBNM3wPuTTop1mjeEJEMqcpCBXObkAIlrw8umSDovOkev8AeEFiuEE1SDXIcoGDi1RplWQQlxP4xszxYGBIFLEbARrAMVNffnuQG7TOfcES3eUPeV+Qv9AjrERriO7JpaM2mZGYHsBOA1hiZyl3KHAovEhL1H/hNximEpgMUaxg5QGKHz3hmV1lfeUeQ/rKd3qOcsnm4kgvhl/7gnzh90g5zpvHiUNvUY5A+YfVVSOlmmaGKmvvz7ILxJ7RVlOUjDYAk4+CpYfsF5a9O1chWUIOnqGKC0hM/bCwdGrroh+A3bLu1Q74E8zFTTnf/wDXjTTop1mjeESw9hMmhupUvjFKJbwZPl0uZM474T6DFAZl6ZI94ONdRsNjxDGApRMPMET+YjNJ25cc5RNYJ/QIpEw4MlkqbfximKh/qMVNPfnuQG4IgUoiPMEUlmYzWeLrfwEGwT+gcShV9TX6G8IesG8xbmQdJAoQeuKQUIcS6yXYV12/Y/iLxqJzQZXPkT8ySw72pAerp3eo5yyebiSC+GX/ALgnzh90g5zpvHiUNvUY5A+YfVVSOlmmaGKmvvz7ILxaoMpBeXFfk9ohz5MSR+aWzhs5LzFOAG+kFMByAYOYQr7lUO+BPMxU053/AP14006KdZo3hDP3pDOF8Yqhy3fpcg+Jzo8h8kYoRM/QJ+QhuRNx+AeJTGZ8GyBbCrf6ZIotLOE5+glzpkGzP9AiqR0o0zMVNPfnuQG5SN56BIHi3xBMQD6jAcShV9TX6G8N2f0KbTSyXaVkHX9jQ/lrqVuRQeJCQwaB4lcQ5Q5wiRO/TpK0X7aYerp3eo5yyebiSC+GX/uCfOH3SDnOm8eJQ29RjkD5h9VVI6WaZoYqa+/PsgvFnDcHUndpD/EkaOYPpEjX9IkrRXrSLuVQ74E8zFA5qzlgvPTFyJWdjWsotrkveKMW1yXvFGLa5L3gjDGbsZmJwZOSLCTnsdyadFOs0bwhp7yhnC+MPmZH8tVbH5lCVoUIoxdmIP4VUD/3CJLMCzOUt3Qfxk5fr8d2n8z9Km5GhPyNy8uUMVOpZvMvWfH51xsS5IRVI6Xa5qKmvvz3IDcqgq2FHLHtql4tCr6mv0N4cSYytrNG4ou0inKOkIn9DHUpEyzWuu1/uXiUDVFSjCIdg5i+rp3eo5yyebiSC+GX/uCfOH3SDnOm8eJQ29RjkD5h9VVI6WaZoYqa+/PsgvFde6LZA+EG/Mf6jFFr2WGaDcqh3wJ5mGMqeTOz9DbnWsOetFqk57vVi1Sc93qxapOe71YoFKH0sXeC8bGRA4F3Jp0U6zRvCGnvKGcL4wT8ofSKey30Od+kk/I580VOJlXIvLz5wm4+dkYsVnKnIVMgmgRWmsz61XKviMS5oRgwRbJ/lTIBYqkdLtc1FTT357kBuVRCV5AQ3ZWLxaFX1Nfobw4s8pEykqA7+cDqjzJBzjExeA/fKOARIgBx/ITm3aAXskzhuII1ghxSiTtTiRV+iBtbwi3GRd4J6potykXeCeqaLcpF3gnqmi3KRd4J6potykXeCeqaKW0jlcxo8ug0eEVVMJaxeJJ10205ZrLGsE01imMbqCvFuMj7wJqmi3GRd4J6potxkXeCeqaLcZF3gnqmi3GRd4J6potxkXeCeqaLcpF3gTVNFuMi7wT1TRbjIu8E9U0W4yLvAmqaLcpHj5NU0W4yLvBPVNFuMi7wT1TRbjIu8E9U0W4yLvBPVNFuMi7wT1TRbjIu8E9U0W4yLvBPVNFuMi7wT1TRbjIu8E9U0W4yLvBPVNFuMi7wT1TRbjIu8E9U0W4yLvBPVNFuMi7wT1TRbjIu8E9U0W4yLvBPVNFuMi7wT1TRbjIu8E9U0W4SPvBPVNDV82epb41XTWJ1kNX/AFL7pBznTePEobeoxyB8w+qqkdLNM0MVNffn2QXizpwDWTO1h/hSGOcPrElQ9HkzRLspF3Kod8CeZippzv8A/rxpp0U6zRvCGnvKGcL4wT8ofSKbSzhCQKGKFdVD8ZYkExGWTps5/gA1Y/0GCHA5QMXlAQrhFUOZbxLU2JOdca5skIoBLfS5yZ0cPwNvMO5VI6Wa5qKmvvz3IDcpm1F1Rh0BecgAfRxaFX1Nfobw3VFCJEE5zAUoc4jFIaegSzbSnlN8VoWWUcKmVXUMoobnMYeJRRqLSjbMg9iy08SmdJ1nbxSXtFBI2S5DjhDfpAARGsHKMS+pyKrQDvnZklTfwEifSJeQvQRWNZkPykUD4h+kl8xcyt0VdmqJDhoN/QYkc2TnMrSdJ8gm5Dl7JviH6h90g5zpvHiUNvUY5A+YfVVSOlmmaGKmvvz7ILxaoU2KiwJL0/aLcp8mJEwNMp02bF+JwE30CClApQKHMG5VDvgTzMVNOd//ANeNNOinWaN4Q095QzhfGCflD6QdMFEzEMFcDBWGJ1Lxlk3ctB5im/D9PhFDZnwjIErMa6iP+meKVzPhOfuFOdJId7J/xFC5ZwdIEsIt+M25VI6Wa5qKmnvz3IDcXRKugokf8pyiUYfszy+YLtT86ZxDiUKvqa/Q3huTiespKgJ3Sn4/4Uw/MaJ7Sl7OziUwik2+CReLKWJpjNG7Un8Zwr/SEkwSSIQvMUKwbqprFE5uoojBjCc5jm5zDXH9I2W9HdJLVrIEzget11hhi/bzBoRw1UKchgiqDNEHjts1QMBxQshOYv6WpocRZPifAqhR/UPukHOdN48Sht6jHIHzD6qqR0s0zQxU19+fZBeJPqVs5MkYoHBVz8Eyw9erzB4o5cms1FBigtHjMGovnRay63MHZLu1Q74E8zFTTnf/APXjTTop1mjeENPeUM4Xxgn5Q+m5VGlvKhMCZs8SCfHkyL4mHTrEyokTA00nLdtzgY9kf6fGCEAhAKUKwAFYNyqR0s1zUVNPfnuQG7VCknNNEMlbiULMBaUNjD1G8IpDTtJrZtpXWVV+KvwLDlys8XMs5UMqobnEeNU+kgpJnma5eU/IlxF/YKZIwG6VJQxa5SHEOsCxvCuCU1BjeFcEpqDG8K4JTUGN4VwSmoMbwrglNQY3hXBKagxvCuCU1BjeFcEpqDG8K4JTUGN4VwSmoMbwrglNQY3hXBKagxvCuCU1BjeFcEpqDG8K4JTUGN4VwSmoMbwrglNQYKRwSvYEWLX56wDG8K4JTUGN4VwSmoMbwrglNQY3hXBKagxvCuCU1BjeFcEpqDG8K4JTUGN4VwSmoMbwrglNQY3hXBKagxvCuCU1BjeFcEpqDG8K4JTUGN4VwSmoMbwrglNQY3hXBKagxvCuCU1BjeFcEpqDAgICICFYd2pn7tMMsn6h90g5zpvHiUNvUY5A+YfVVSOlWmaGJDP1pAqsogkRQVQABsoukP8AFEIPVFmZvyIIEh7SybviiCjsSF6k/wAMNmjl+vYNklFlDdQV4o1QYGhyupnWOrzlS+BYAKwbtUO+BPMxI6ROZDv3oxEz77Wr2cXRpngG0XRpngG0XRZngG0SOm7+ZTlBosigBFB3Jp0U6zRvCGnvKGcL4wT8gfTcn0uCaSZw2+Ji/h+sCUSGMU4VjFGsMVOZbyLzA+bJu1SOlmuaipp789yA3V0E3KB0VSgZM4VhAYpJR9WRPh+LU/KmfdKYSjXKIgP9OPRejqk8e1zhWaJ/nN1/0hFIiCRU0wApChWAA4i/sFMkYDdoHeohln8w7lYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFYIrBFaKRXxzHPm3amfu0wyyfqH3SDnOm8eJQ29RjkD5h9VPaKtp84TWXVVIKZbH8EXOJfjK8XN5fjK8FqcS3DuIbUGkzbnQFXOGhszbs07BuiRMvUUK3FndEWs8eA5XWVIYC2NYsXOJfjK8XN5fjK8XN5fjK8XN5fjK8S2gzKWP0naS6xjpbjhEHDdREwiAHKJRhOp0wSOQ4Lr1yjAcgbrygLB48VXFZYgqmshKWJZLkpVL02iH5Ew3Z7RVrPnKay6qpBIWx/DEhou2kCqqiCqpxU7fEfy9vMmh27ogHTNFIKJOpKcyiQCs07Yc5fUUdoe6m5yrLgKDT+54ZMUGDUjdsQCJl5gDir+wUyRgN2gd6iGWfzD8lpFfJMc+bdqZ+7TDLJ+ofdIOc6bx4lDb1GOQPmH5UYoHKJTAAgPOAxN6CMH4mVa12qw9nmh9QmbsuZEHBOtOFWLpAayzZYg/wBSDFgfsH1YRYO3A1kWqxx/oQYYUHm7zlUSBuTrUiT0HYS2sqvXcr9Z+aAKBQrAFYOMv7BTJGA3aB3qIZZ/MPyWkV8kxz5t2pn7tMMsn6h90g5zpvHiUNvUY5A+YflolKPOADHo6WDJqwBQDmAA9Sv7BTJGA3aB3qIZZ/MPyWkV8kxz5t2pn7tMMsn6h90g5zpvHiUNvUY5A+YfnC/sFMkYDdoHeohln8w/JaRXyTHPm3amfu0wyyfqH3SDnOm8eJQ29RjkD5h+cL+wUyRgN2gd6iGWfzD8lpFfJMc+bdqZ+7TDLJ+ofdIOc6bx4lDb1GOQPmH5wv7BTJGA3aB3qIZZ/MPyWkV8kxz5t2pn7tMMsn6h90g5zpvHiUNvUY5A+YfnC/sFMkYDdoHeohln8w/qZ3TFhJzCl7wvgyQ9p7NnPsBI2J1F+4wNJZwYekXGtDWmU6bD73vucKAxKKoaC9ZOZo7wbCE5SwismumVRI5TkMFcDFGuA8ekV8kxz5t2pn7tMMsn6h90g5zpvHiUNvUY5A+YfnC/sFMkYDdoHeohln8w8ScTlvJGgOHQHEgmsPwBXGLoko7DnUi6HKOy51IZ0ykz0bArsEz9SoWMEOU5QMUQEB5hDcMYClEw8wBXi6HKOw51IuiSjsudnF0SUdlzs4kk/aT0ipmgKVkhADWZYfvUpcxWdLV97SLZGseeCVQZQc4FrL6kAICFcOYd1+9RlzJV04GsmkFcYJVBlJzgQCOdSA3KY0uMgc8ulx6x+ZVUPAI+IiPKPFo1SZeRrgQ4iozN+dP/ANCGzhJ23IugYDpnCuUQ41Ir5JjnzbtTP3aYZZP1D7pBznTePEobeoxyB8w/OF/YKZIwG7QO9RDLP5h4lUW99PPlih1GZdOJQdd4kY6gKiSLQpJgFNqMTGpwnYGNLXJ8hWJRPZhReYi1dgcUSjWURN4hDZwm6bEXRMBk1AsijC/sFMkYoxLkJrPyNXQCKRgP/aLQZLgldqMWgyXBK7UYlEjZyMihWRTlBQQE1kavFLL1n+agGZjys7svMRYEz/0rhyRQ+ZcJ0fQMYa6iX+kf/jdqjTKwaoMCc6g2Z8kIXZmYTFuif89ZM5v6CPLAc0UpnHA0lUWJ7Y/4E8oYEwmMJjDXEeUR69yh0lazs75B18EyiQ3xLE9ou9kZxMcN+bfBYoePEqeTkSqnla2Wj9uNSK+SY5827Uz92mGWT9Q+6Qc503jxKG3qMcgfMPzhf2CmSMBu0DvUQyz+YeJVFvfTz5Yqc3vq5827VDlRVpcSYEL/AKqJgKfJGKnL8VpauzOPsDVyZIwt7BTJGKDX2JZJ+JSy9d/moofL+FJDOWnbAtjlRU+mItJwsxV5l/OXcEawVxi+mnXWgCn/AOssUqv0Wy0/AIDmCKozwTzJs0wSdnp3amvSD7Nlg6ZFUxIoUDFNyCAhFIaBAaycyfYfaFUlEFTJrEMmoTkEpgrCG5KXYsZs1cYNUNEByhX4tIr5JjnzbtF6Tko8m5KdsZYVjFHkPF0xLu0+1i6Yl3afaxdMS7tPtYumJd2n2sXTEu7T7WLpiXdp9rF0xLu0+1i6Yl3afaxdMS7tPtYumJd2n2sXS0e7lNrEkpMwnnI3OJVg50j8hvWPukHOdN48Sht6jHIHzD84X9gpkjAbtA71EMs/mHiVRb308+WKnN76ufNu0qIB6MP80MVNul3WYhb2CmSMSx66YTEF2AV1wsgD8Fl/aLb6TYMdhFt9JsGOwijD13MJImu/CsuJjAP4LGKV3rTDNRUz9lMMokUpbHkVLfS0eQpzAuT/ANCGbkjtmk4TGuRQgGCKXTLgyQLnLyKK/wCkT6jFTmWWDVd+fnUGwJ9Aild+i+cT8AgOYIp3fStmybtTXpB9my7tKpZKXTIy8yOVAxeZYPzf/wBgwFA5gIayLX5BrVq+42r+jJ1+ewDi0ivkmOfN69q6VZOk3KBhKomayKMMXQPGCDgvMqQp9IerfdIOc6bx4lDb1GOQPmH5wv7BTJGA3aB3qIZZ/MPEqi3vp58sVOb31c+bdps7BtRhyHxWrJlipo3EV3znqKVOFvYKZIxQa+xLJPxKWXrv81FTP2UwyiRT+W+lyUHRArna+UYqeTL0mUHZnH8bY3JkjFPXpn87QlqHLvXnNEqYklstQak5kiAH3ilV+i+cT8AgOYIqjNBTmyDr4Kp2Ojdqa9IPs2Xcn9MmkoskUKy7rsfAv1iZTR3NnIrPFROb4B8C/QNyWNjPZo2bhzqKlD+8FCsWsHFpFfJMc+bjyigaMylSDsz1QgrEsqwEi5qh3grswi5oh3irs4WqcIJInPwgr+EuD41Hb3WGYJ4erfdIOc6bx4lDb1GOQPmH5wv7BTJGA3aB3qIZZ/MPEqi3vp58sVOb31c+bcdO0GaBlnCpE0y85jDFKJ+ekUxIi0KYUCDWSL8Tm64oxJ+BZMmgf2xvxqZQwt7BTJGKDX2pZJ+JSy9Z/moqZ+ymGUSHCJHLZRFQK5FCiUYkzw1FqUKkX9mSySU+nwGKHtTzqlS0wX5khFUcoebcpXfovnE/AIDmCKXSgZvJFCJcq6P4092gDxBivMF3ShU0ypFikFOl31khLK6CGE/jNHWO7U9k4rOzzNUv4EvwJZXGpFfJMc+bjs6YTZizTbILJgkmFiWumEW9zvDpbIIt8nmHS2QQenM6UIYhl0tkHGo7e6wzBPD1b7pBznTePEYUvmstZJtWyiYJJ8ha6cW+TvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KLfZ3hkdlFvs7wyOyi32d4ZHZRb7O8Mjsot9neGR2UW+zvDI7KDU6nRi1hVR2XEoHeohln8w8SnjRd5IyEaonVPvwchIlylJpUgKDJs6TTE1lW3iBmVMVfg82EEorSKcKgZ5Z5ThSKP0PaSSsscd/dYT7BuLAIonAOyMNGE8l7v0hoydpqhXrGBKOE6YdTzYRwnTDqebCKHOJsug6GcAtZAYLDfCWMUmRUXo2+SRIY6hk6xSlip6xdMiPfSkFUbMS1rMu5TqQuVpom8ZN1Fd9JWOCZfiEUKlB5XI66xBIuuazOHgG5SSVP16WrLIs1zpCclY5SDWgObcplRI5FDzKXErkNyrJF8Q41H6POJ67ApAEjYvtFftDJoiwaJtkCgRNMKwBxqRXyTHPm3ZBQMi7UjmamP+PlKiTk0xaJI8XU2potDkeAU2potDkeAU2potDkeAU2potDkeAU2potDkeAU2potDkeAU2potDkeAU2potDkeAU2potDkmAPtTQ9qey1VEQaGVQU1gh+wXlr1Vq5LWUTGsP3ijt7svzBPD1b7pBznTeP8gUDvUQyz+Yf1M8oOymZjLNh9FXHVND2hc5Z8zbfy9pL7QMomJRrCwdbIYbUYnDvkTYLB/U/4PGJRU75lJquGaS+8NmqLNAqLdMqaZeYpePSK+SY5825LyFVmTUh+UpliAOmA5vWVRiAE8QN2kOXWijt7svzBPD1b7pBznTeP8gUDvUQyz+YfktIr5Jjnzbkr6XZ58nm9bVI6ZbZj/KKO3usMwTw9W+6Qc503j/IFA71EMs/mH5LSK+SY5825K+l2efJ5vW1SOmW2Y/yijt7rDME8PVzRMUps8J2Vzh/cf5AoHeohln8w/JaRXyTHPm3JX0uzz5PN62qR0y2zH+UUdvdYZgnh6umrIWdJl+wvWVL/IFA71EMs/mH5LSK+SY5825K+l2efJ5uLPKcM5YcyDUPSVw1Sw6pvOXPMuVAvUmWEqXztIffjmygCJVVFPyEmiAZ1L7Q0eIPm5V2yhVEz8xg4lUjpltmP8oo7e6wzBPD1dPZOL6VA7RLXVa+T+QKB3qIZZ/MPyWkV8kxz5tyV9Ls8+TzcSnFJTtA4NZHrKmCuqfshuCycAyB3vRvRxMJbMOavu0apCrI33W1PyKE/wDQhJQiyRVEzAYhwrgIfHdqkdMtsx/lFHb3WGYJ4erMUDAIG5QHnCKW0aPJnYroFrslR1B6vn9A71EMs/mH5LSK+SY5825K+l2efJ5t1dYqDdRU35SFEw/8Q8dHevV3KnKZU4mHcoEimvRYyapAOQyp4pDQLncyfYfaDkOkoZNQokOXkEohyhuUAfi7kQoH52x7DdqkdMtsx/lFHb3WGYJ4escNknaB0VyFUTOFYxR+MT+gzhkJl5YAroYP+MkCAlEQEBAQ5wH57QW9VDLP5vkleKRXxzHPm3JX0uzz5PNuz+999Y4A3hAc25U8vcHPG3KYS2TqsxXfnBBf+BQv5hjwipnzzAMjdqkdMtsx/lFHb3WGYJ4eumVHZbNuV02IJ8IXkNDmpsgb3V8qTLIBouaOcfR1IuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYuausfR1Bi5q6x9HUGLmrrH0dQYPU4dEIYwv0tTiIv3aCdgg7XTJ2SKCARwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjRwtMMfdbY0cLTDH3W2NHC0wx91tjQc51TmOcxjnMNcREa4juSvpdnnyebdcog5aqom5lCCUf+YctztHSrdQKx0jCUdyp7e2OeNFIabNpZZoMqy7oNUkPn7mZORXdqmUUH+303KnrAW0kO4PzuT/2DdqkdMtsx/lFHb3WGYJ4fyOv7BTJGA+VSvpdnnyebiU6o2cxxmjQmfKHm3Ep69QlHByCm9IiYTGEv5jV92j8jWnkwKiQKyJeVU/UEN0CNm5EUi2JCFApQ3apHTLbMf5RR291hmCeH8jr+wUyRgGq+AV1Bj0VfAK6gx6KvgFdQY9FXwCuoMeir4BXUGPRV8ArqDHoq+AV1Bj0VfAK6gx6KvgFdQY9FXwCuoMeir4BXUGPRV8ArqDHoq+AV1Bj0VfAK6gx6KvgFdQY9FXwCuoMeir4BXUGPRV8ArqDHoq+AV1Bj0VfAK6gx6KvgFdQY9FXwCuoMeir4BXUGPRV8ArqDHoq+AV1Bj0VfAK6gx6KvgFdQY9FXwCuoMeir4BXUGPRV8ArqDHoq+AV1Bj0VfAK6gx6KvgFdQY9FXwCuoMeir4BXUGPRV8ArqDHoq+AV1Bj0VfAK6gx6KvgFdQY9FXwCuoMeir4BXUGPRV8ArqDEsKJZw0AQEBBcnm4ggBgrDyhE9oEi8OdeWnBBUecn8Aw6opOWhhA7FQ+b/FCVHpsqNYkucalaJVU9drmA8xUBBPsE5TRL5c2lbQrdomBCBxKpHTLbMf5RR291hmCeHyQxgKUTDzBywpVEmm+nsE21hkj94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvF0SbYNrqD94uiTbBtdQfvDKn8xVfIEXI2BI6hSnrFHm0/OX1/Z/35fNAesqkdMtsx/lFHb3WGYJ4fJKRuvQqPvlupIQD6jyfpeblDniVOfTZW2cYRIpv7cYTAHOIRZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYsy9oNMWZe0GmLMvaDTFmXtBpizL2g0xZl7QaYe39n/fl80B6yqR0y2zH+UUdvdYZgnh8kqiOt6kSaGHVDQH6agTr0ijKZMAcyfGp+spbKJfgVEsWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMWZ+2bTFmftm0xZn7ZtMb4ftm0xK+l2efJ5oD1lUjpltmP8AKKO3usMwTw+SVR3VnM2rbBJ2en9NU1dcr5rkqBxqokoOfeZklzECwV+ZS5QqMyaqHGsUqxBH6V4AQEK4esqiqFPPUSdhCKO3usMwTw+Qzmes5G2310blN+RMvKY0L1R3pjj6M0QIT/8AOuYYujTXANYmszWm787tyBQUOABWJzBW/TSecLyR6LlqBDHEgkEDxdFmuAa/3hrVIdgcPS2iJyf7cSmbtZy139oeyDmMX4lHdUSIsmZNQoGIYKwgPMMTqp6NmZaUHDMH/wDBh5J37Awg6aLJ/wBRLyafmMgp4Zg2K1mKR1iE5CqE54uiyrBOtSLo0qwTrUi6NKsC61IujSrBOtSLo0qwTrUi6LKcE61IuiSjButSLospwbrUi6LKcG61IuiynButSHtUdoCA+gtljq/DfOQsO3az50o5cns1VBrmGKO3uy/ME8PkNJpieZT5yqY1chDimTJD9ZQ2YnYUhQL/AALjvRw8OKIAMKytkvyrNED5SYDFrko7ta7IItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItak/drXZhFrUn7ta7MItbk/drXZhE9STQnz5JIgEIRYwFKHMAcajFDxnSAunRzJNuYoF5zxNaniANTHlqyu+l/gU5QNAgJREBCsYOQQ41Hb3ZfmCeHyGYpCjM3aY85Fjh/f9ZR5MVaQsChhy+PzWkV8cxz5uNQl2k5o02Inzo1yHLCqpEUjqKGApChXERh+sRzMXKyfIRRUxi/QR41Hb3WGYJ4fIacSwzGfqLh7Fz+MMr4h+soBLDOZwZ6b2TYukw/Nacyk7KdHdgX/Qc+bjMJk7liwqs1zIn+NbmH6xMKRzOZpb07dmMn8SFrFAdHGYMVpi9TatyiZRQdH9YZtitGaKBPypEAgf8fIZlK2s2aCg8TA5B0gPWEXOJbjLqLm8uxp1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxp1pL9oLU4luMOoYMG8ualbtEwTTL81dtEXqBkHKRVUj85TBC1TuVnPXTVcpB2QN9wi5vLsadaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxt1pL9ouby7G3Wkv2i5vLsbdaS/aLm8uxl1pL9olMhYSYglZo1jDznNymH/wCFf//EACwRAAEDAwMDBQEBAQEBAQEAAAABAgMEERQSITEQIDMTMDJBUEBgIlFCYSP/2gAIAQIBAQgA7duu5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5ubm5v8AvolyOkc4bQtQw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMWiavElLI3j9uNiyLZIaZI/wZJkjQzluRyJIl0/AmpUfuParFsv7DWqq6SCJI2/g8KV19ulDt+DZVKiFHpt+xRx6naj/8/BvuSxJIhg7jI0jbZPwquPS+/wCxSpojv2XLoXQuhdC6F0LoXQuhdC6F0LoXQuhdC6F0LoXQuhdC6F0LoXQuhdC6Gy9b3LoXQuhdC6F0LoXQuhdC6F0LoXQuhdC6F0LoXQuhdC6F0LoXQuhdC6F0L9lY27LnH68fHZUPX1FNamtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3Gtxrca3GtxrcUrl9QXrP41/Yj47KjyL/gKXyJ2T+Nf2I+Oyo8i/wCApPKfXWfxr+xHx2VHkX/AUnlPrrP41/Yj47KjyL/gKTyn11n8a/sR8dlR5F/wFJ5T66z+Nf2I+Oyo8i/4Ck8p9dZ/Gv7EfHZUeRf8BSeU+us/jX9iPjsqPIv+ApPKfXWfxr+xHx2VHkX/AAFJ5T66z+Nf2I+Oyo8i/wCApPKfXWfxr+xHx2VHkX8trFXhlG9eUompziRi0LB9G9o5jm8+9SeU+us/jX9iPjsqPIv8m/8AJx0gpdW7mxpH8e10TXk9Krd2+7SeU+us/jX9iPjsqPIvtRQufxhSGHIYcgtLIgrXN57IoHSJthPMN5hvMN6j6Z7E3ES5iPJad7E7mN1LYw3ksDo03601Pfdbd3JZTnYqqdE3b7lJ5T66z+Nf2I+Oyo8i+1RfFSWZI0S+YwzGDaiNeFYjyakVE1NXrR+MkmSO18xhmRmYwnqWPbZPsj+RezbkjdTLDk0r20TN9ZfdSu4TrSxa3XLdZX6GqpHK16d9RFod7lJ5T66z+Nf2I+Oyo8i+1RcKV3CdkdQ9gyTWm1VCjF1J0o/gV312/ZH8kH/Agk1tKuPS7sTmxC302WIX63KV+yJ1p2aGH1frVbxqMe5q7Q1erZfl21LNbL+5SeU+us/jX9iPjsqPIvtUPCldwnbRyWXSTN1M0lrdKLdhXfXb9kfyQd8SjfZ1iqZdnZTM1vKh+hhRcKV3CdIW6noJ2VXjXrDVKwikbIm33v0tySNs9fbpPKfXWfxr+xHx2VHkX2qHhSu4TtiXS9FOUJdnL0o/gVkTnWtjyGPIJTyDmObz9kfzQd8RrtLrif8AbSVmh6p1o2aG6itfd2koeFK7hBCj3k7ar4L2Mkcxbtgqkfs7rVJaT26Tyn11n8a/sR8dlR5F9qh4UruE7WcoN+JP5F6UXjJJUZa+RGZEZkxlXIx/H2R/JBfiLyUb9TdJWs31dGN1LYT/APmwe7U65Q/FSu4TpR+TtqvgvbDC567RtsnWr8nXnokEimNIY0hjSGNIY0hTQvZJddldv0mbqYqJjy3MeQx5DHkMeQx5DHkMeQx5DHlFppDHkMeQx5DHkMeQx5DHkMeQx5DHkMeQx5DHkMeQx5DHkMeQx5DHlHMc3m/9EfHZUeRfaoeFK7hO2NLuQt/ySrd69KPZhXfXb9kfyQX4n2Uz9LydmqPpRM1OuVklkt0ofipXcJ0pnWk7ar4L1RLkFL9va1G8CH/pUOvIvZS09k1KidyKbdm6exsPrUvtDMkifx7j4mv5lZodb+iPjsqPIvtUPCldwnbRx3dqJn6GKvWj+BXfXb9kfyQX4i8ibLchdrZcqWaXlPHoYhUv1O6UPxUruE6NWykSo9NXZVJZi9I4XSKRU6R8dkq6G3F3W/VvInH8aDk2HsVq2WjjVm6/yV3Kf0R8dlR5F9qh4UruE7IqdzxjUjZYqptTtKdKP4Fd9dv2R/JBfiLz0on76SWFJHITP0sF36UPxUruE60k1v8AnsqUVWKQ0qu3VrEb8b9tXLddPY3kTjrqQ1IakNSGpDUhqQ1IakNSGpDUhqQ1IakNSGpDUguhxqaakNSGpDUhqQ1IakNSGpDUhqQ1IakNSGpDUhqQv1rvr+iPjsqPIvtUKIqKTwpLZDCaYTRlMxormsJ6pXbJfrRWVpNC2UwmmE0wmoTUrY236R/JB2zReVEIXaHXG2duVkl9utD8VK7hOrXK1bpBMkqWUTvuqE8yMQVb9jeROOtX5F/Fg8ada7lv9EfHZUeRfahqFjRURK1yGa4zXC1T1HOV3PZFUrG2yJWOQzXGa4zXKSVbpG26NWyi1arsXv1bVualh7lct+sU6xoSzrIm/VjlYt0hqUf7E9UjNh7let17G8icdavyKJ+JB4k613Lf6I+Oyo8i/lXsRVbmjKuNT1GKakFkRB1WxvElU53Hc3kTjrV+RRPxIPEnWu+v6I+Oyo8i/m3NSl/ZbyJx1q/Ion4kHiTrXfX9EfHZUeRf2G8icdavyKJ+JB4k6131/RHx2VHkX9hvInHWr8iifiQeJOtd9f0R8dlR5F/YbyJx1q/Ion4kHiTrXfX9EfHZUeRf2G8icdavyKJ/RFTOeMo2oY8aDqVi8SUTk4VLd8HiTrXfX9EfHZUeRf2G8icdavyKJ1ZGr1smHIYciDqeRDdORDEkUw5DDkJYnR/JrdTrJiOPu3VjVeukxHp1pabVu7/8ERFLr1np0kQc3Sule2DxJ1rvr+iPjsqPIv7DeROOtX5FE60fkKmoex9kSrlGVi//AFJE2Zt0VqtWys+RPI6OPbLkMuQkldJ8qfyIak1aSoZpkXrRRqqq4R6uYq9KePW4t/5wVUixrtFUpIluyti/+k7YPEnWu+v6I+Oyo8i/sN5E461fkUTrR+QrPJ1o5LO0lYyztQ3kqt4uyn8qFQ7Q9rirZqYj+nOx4oSn8XShZ/yq9a76EWykFZ9Pa5ruF5Jm3aL2weJOtRTLIpgqYKmCpgqYKmCpgqYKmCpgqYKkkCs9yPjsqPIv7DeROOtX5FE60fkKzydafyoV3CDeRzEdHvjwGPATNRjtqfyIV/DSB3qw6RzdLrFMzXIVj90aQeE+yj8fWu4TrA97V/5QXgdyvbB4k62NvbkYj2j26XW9uPjsqPIv7DeROOtX5FE60fkKzydaZt5Cudwg3kqfF2U/kQrvoo36XlYzS/UUjdLNayO1uVSDwJ0oXf8AKp1ruE6Q06v5jiSNNl/9JnWbfug8Sd8lYrFsZqoZyja1RO2byL7cfHZUeRf2G8icdavyKJ1o/IVnk6NaqlNGkSXWok9R43kqfF2U/kQruUGrpdcez14yod6cehEKfeEUp5ND+tYxX20wUlt37fR/+FZJb/jug8Sd76WN63MOMw4xKSPum8i+3Hx2Ppo3LcxI0MOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMOMw4zDjMSNFOE26VfkUTrSuRr7q/0HrddFOhkQs4lqFkEG7OFfC5ll0U5opyqbGipogVElRVq3I5diln0JZaiTW7pFK1IrdaWoRNnc7lixY3Jp0iSyOerluvbT+NBRdkJqtb2blymXIZchlyGXIZchlyGXIZchlSDayRBknqMJdnr7cfHYuwq3/Xs3sq/IonW+38iEVW5mzm1UanqMUdURt4fWu/8Alzlct17oPEnSTj3aJf8AhSb5r7cfH+Aq/Ion4kHiQ+iT4e7RfBSb5r7cfH+Aq/Ion4kHiQ+iT4e7RfBSb5r7ce7EF/fq/Ion4kHiQ+iT4d+3dRfBSb5r7dI68f8AgKvyKJ+JB4kPok+HbFSudy2kYnK0sQ+j/wDHNVOyi+Ck3zX26WXQ7Sc/v1fkUT8SDxIfRJ8Oylgv/wBL9mtqOsW/944qIdaXFSy260XwUm+a+2i2KafUmlf3qvyr+LB4kPok+HVu7rESIxtj7KpVSQhq77PRyLxfaxVs0uv1ovgpN819xrrbpDVI7Z92r8dhVNzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Nzc3Lr2VS3kPv8AD3KfxofRJ8OsXyQ4ROlZ5OlM997IV/11ovgpN819y3RkzmcNrXGeZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZxnGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGahmoZqGaglaiqIt06uYjtz0kPSYekw9Jh6TD0mHpMPSYekw9Jh6TD0mHpMPSYekw9Jh6TD0mHpMPSYekw9Jh6TD0mHpMPSYekw9Jh6TD0mHpMPSYekw9Jh6TD0m2ESzT6JPh1atlQY7UidKvyEFKr93Mja1LN+isfdyJ1ovgpN81/wAO3kRNvybbEnw7KWfTsp6LHOuuxsTypG2w5yqt+tF8FJvmv+HbyI9LGtDWhrQ1oa0NaGtDWhrQ1oa0NaGtDWhrQ1oa0NaGtDWhrQ1oa0NaGtDWhrQ1oa0NaGtDWhrQ1oa0NaGtDWhrQ1oa0sSfDsTYhq1b8m1MbhaliEla1OHvWRbr1ovgpN81/ETcSjaYTDCYYTDCYYTDCYYTDCYYTDCYYTDCYYTDCYYTDCYYTDCYYTB1GjUX9lng92i+Ck3zX8SFt3Igljb+NUVyKSJpcqd1iyllLKWUspZSyllLKWUspZSyllLKWUspZSyllLKWUspZSyllLKWUspZSyllLKWUspZSyllLKWUspZSyllLKWUspZSyjPB7tF8FJvmv4lGn/9L/zVaWlVe6jaisNKGlDShpQ0oaUNKGlDShpQ0oaUNKGlDShpQ0oaUNKGlDShpQ0oaUNKGlDShpQ0oaUNKGlDShpQ0oaUNKGlDShpQ0oaUNKGlDShpQ0oaUNKEnHu0XwUm+a/iUTf+VX+aubwvdRyWdb9KRNvdovgpN81/BjiWRRtF/7hNIo2xpb+X7JoUkQwmjqJo+JWLv0RbEVZbZzZmP4unS5cuXLly5cuXLly5cuXLly5cuXLly5cuXLly5cuXLly5cuXLly5cvY3J6RFW7cR5hyGHIYUhhSGG8w5DDkMOQw5BtE/7YzSlibyL+DTRo1pe/Tb+Xk3NjYqYkczV27px6jz1pD15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D15D133IHOViKuxq6ohPU+mtkZVuVd0cqt7LL0m8i/gx7sPr+uo2jX9Wn3jTuqWqj1UaiuXaNqo3t+ibyL+DRvuz+ysk/5sfX6f0Ukn/Nu3hB8TXt3bTsbxv9dkj9LBy6lv8AgxvczjNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzNeZrzMeOcruf1GusJWPQzXma8zXma8zXma8zXma8zXma8zHj5XP5T8Hfpub/z7m5x/gNzf2//EACERAQEAAAYCAwEAAAAAAAAAAAEAIDFgcICQEFAwQEGg/9oACAECAQk/ANiCIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiNg89gc+HzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzMzM/2o5cBiI+L91oz6MiIiIwkRERERERERERERERERERqwjh6+X0memSPDjIiPs567eqo5GszMzMzMzvyREa9IiNdszMzM6qZ7p2IiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIiIwkRERERERERERERERERERERER2nOw7MzMzMzMzMzMzMzMzMzMzMzMzMzMzMzPQx//8QAMREAAgEDAgUEAQQCAwEBAQAAAAECAwQRFDEQEhMgMxUhMlAwBSJAYEFRI0JhUnBx/9oACAEDAQEIAOz27MmTJkyZMmTJkyZMmTJkyZMmTJkyZMmTJkyZMmTJkyZMmTJkyZMmTJkyZMmTJkyZMmTJkyZ/oLaiVbyECV/M1tQ1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tY1tYjfzW9O7hLdPP3dSoqay6905/QNlGi6g7BOJVpunLD+gt7p0/Z05qosr7icklzFxWdSQx/QWDXD9Qw/ocotq7psTyvuL6ryx5T/36FbFGs6bFf8AsTqOo8/R2VTmhy/cXb56g9sccMwzDMMwzDMMwzDMMwzDMMwzDMMwzDMMwzDMMwzDMMwzDMMwzDMNcGhLBhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZh9lnPln9uyrv2W1OLpoVKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKmdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oHSgdKB0oF3BKmxr3423kX272Ku/ZaY6S/oF74mY9+Nv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCv6Be+F9lv5EL7Z7FTfstPCvsfb8974X2W/kQvtnsVN+y08K+rlUUSd9BbO+n/jXVBfqEiF9CRCcZfnvfC+y38iF9s9ipv2WnhX8nP8D/APte7x8JVXU+XYyFZwLe8UvaX5b3wvst/IhfbPYqb9lp4V+KtXjTNdA1tM1tMjdUmRlGW3ZVuI02a6BraZraZrYYKV3CbwIlLBrqZSuozfdUnyLJroFGvGpx/wDS6us/tjnuXsPAnj3LS6z7S/Je+F9lv5EL7Z7FTfstPCvxX/yRSoOo2aGZoahK3qwRCc4Mt7xSfLLjfeQo0HUyaGoaGoaGoULScJZf+Cr8GNZeCnLlmmQfNHPbfVMLkMexYcbut04Y7KVPnliM6MoP37U8FrWVSP5L3wvst/IhfbPYqb9lp4V+K/8Akiw3fZXt41ETh02WddzXLLhffM/Tv88MGyP8cKvwZD5FxT5JFlU5odjeEV59SpkrU+RI/T/dvgy5qdSZt7cbP2qLE6cZr3r2nJ7rGO21qOnPH5L3wvst/IhfbPYqb9lp4V+K/wDkj9P3fbe0srmKM+SpkTzwv/mfpz37qvwZH55L2GYcxZ1MT7Lqp04FtT56hfbosN3wrz5abH2WfkXGtaKZUpypvDwse3BPYoyzBfjvfC+y38iF9s9ipv2WnhX4r/5I/T9321lmDRj3KDzHhe/MsqyhnOppmppjuqRCrCexV+DI/PBOHNDlMdORQnzwT430+aXKWMMR5i/3RYbvhfPFPttPIuydKM1h17Nw948bN5p/jvfC+y38iF9s9ipv2WnhX4r/AOSP0/d9s/ixlt4lwvvaZTpynnGmqmmqmmqMs6UoblX4Mj8iOxew5JcxYT/by8Jy5Y5PJUKUeWGC/wDkiw3fC+8fbaeRdtavGmvetLnZsuFn412Z/wBu5po1VM1dI1dE1dE1dEuq9OdPC2j7cKM0pZauqWDV0TV0TV0TV0TV0TV0jV0TV0TV0hXdM1dI1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1dE1NIjOMtv472Km/ZaeFfiv/kj9P3fbWeIMTKKxTXC++Z+nte/dV+DI/IjsXdPngUJuFQT9sl9PlhgsqeXnhf8AyRYbvhdRzT7bTyLi3gr3n+ISk28yNzHuW6xTXH/0u7hy/bFt8PY9j2PYwu3c9j2PY9j2PY9hJMp2DwV6Lps9j2PY9j2PY9j2PY9j2PY9j2PY9j2PYwiFWUChVVRfx3sVN+y08K/Ff/JH6fu+29qpR5VQp888C24X3zP0/wDz3VfgyPyI7EllYK8eSeC1qKcC5qc82WdPlhwv/kiw3fCSyirHpvlHtxtPIuFWtGmitdSqPD7KMXOfKRWI44z2Jb/w2QlhkKkZLJfVFL2X8T9P2f8AHexU37LTwr8V/wDJH6fu+yvdKmTm6k8lnQ5Y5fC++Z+n/wCe6r8GR+RHbhfU/wDsUqzpRZRh1JkViPC/+SLDd8b2jn93ZaPlmiveKO0pyk8ttNdtlRx+/snsyW/HlZys5GcjORnIzkZyM5GcjORnIzkZyM5GcjORnIxdTYUJLfkZyM5GcjORnIzkZyM5GcjORnIzkZyM5GcjORmOP6ftL+O9ipv2WnhX4r/KkihXlSbNfI18yd1UmKEp729pj3aXG+T5ijXlRZr5Gvka6TKN5Kc+XhU2ZH5EduFeHPEnlPBY0se/G/Tyj9Pzl8ZRUlh3NB0nlMffhMtqDmyKx2T2ZLfjZJdJGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhFx5Hx/TtpC/jPYqb9lp4V+KtbKpJN6GLNBA0ECNnTRGEY8N+Na1VWWTQxNBA0EDQRKdnGnPm4SjlCs0jGOMrNSeSnDkWONa2VRlG3VNvsnBTWHXtXD8FvaObyU4KCwuyezJb8bLwr6W58j4/p20hfxnsVN+y08K+qaUirZqZOzqI6c4nIxU2QtJy3pWcI7pY7p7MlvxsvCvpbnyPj+nbSF/GexU37LTwr63CORCS/DPZkt+Nl4V9Lc+R8f07aQv4z2Km/ZaeFfcT2ZLfjZeFfS3PkfH9O2kL+M9ipv2WnhX3E9mS342XhX0tz5Hx/TtpC/jPYqb9lp4V9xPZkt+Nl4V9Lc+R8f07aQv4z2Km/ZaeFfcT2ZLfjZeFfyat0oE72bNTUZG7nEp30X8oyz33PkfH9O2kL+M9ipv2WnhX3E9mS342XhXZUq9NZeugK9psV1TE1L3RnCHeU17Gtga2BRrwq55ZyUE21ewE1jPGc4wXM1eQb43d1y/ti/8AY3gx/vhb3LpsjJSXMu258j4/p20hfxnsVN+y08K+4nsyW/Gy8K7L7xlpbwqQy9DSKlj/APNKtKjLDjLmjlT+JRpxqVcPRUzRUynRhST5bnxM5fbJa1FOnxvqiSUTkUZpCLmpyRG/9v3LSkqixKtaypvL4L2LGt/0fbc+R8f07aQv4z2Km/ZaeFfcT2ZLfjZeFdl98Cw+HG/pZjzFjUzHlJe2S083Zc+JlrDnjJFjPkm4M29zzVy4WKxHYv5+6XGw3Y0mXFmt4SXJuvdFCXJIW3bc+R8ba4VJHqKPUEeoI9QR6gj1BHqCPUEeoI9QPUCjcRqfjexU37LTwr7iezJb8bLwrsvvgWPw43PiZYfJk/iynNwllaiuaiuW8nOGXc+1Jlh/2LiPSrcyg+aOS6qclMsae8y58+BbF75muP6fu+HsXMKcl+5+wtyG3bc+R8cv8lKTjIpy54/iexU37LTwr7iezJb8bLwrsvvgWPw43kuWkyxXu2S+LLTzGEYXC58bP0//ACXtPmgWNTMOUu5c8+RUo8kSv73DFsX0eWfNxsN3wr3SpoqVpVN1sUY80uQW3bc+R99KxU48xoE0aBDsUkPsW5b+NfiexU37LTwr7iezJb8bLwrsvvgWPw4SmkXFV1nhW1LpUyfxZaebsufGyw2kTXNFohPoTZbR6lTmbLj2riLmn1KY+FlUVPm5ri8z8ct7+y22LKln9/dc+R98LqpTjymtqJGuqDvKnaty38a/E9ipv2UrupBYFe1GzW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DW1DWVGvfOX78LLwrsvIuVPCpuvCOFz3LNPXnvRtVT4TWYvHJWhPK6lwdS4LR1Wn1K6bpNKyg47l7buUsq0pckfcrUZOtkW3C7tW/eL/1wyLLOVot7d1XlwgoLC7bnyMQlllCyz8tHRNFSNFSNFSNFSNFSNFSNFSNFSNHTJWNMnT6cyh7wX4nsVN+x4Ekj2MIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjCMIwjMuyy8K7sLs24YRhcfZcMccLsq2cZ+8Z2dRHRqIVtVntSsP/uEFBYXdc+R8Ke4tvyXyxMoeNfiexU3/oFl4V9Lc+R8KfkQvyX3zRQ8a/E9ipu/6BZeFfS3PkfCn5EL8l980UPGvxMqe1R/0Cy8K+lufI+FPyIX5L75ooeNfjvIYqH+Pv7Lwr6W58j4U/Ihdla7jHZ3k3srqqt6d9/uE0+y++aKHjX47yjzrmMY+/svCvpbnyPhT8iFxu7jH7UKm3AX7hP/AHbV+nMjLm9+N980UPGvxtZLu35HzL76y8K+lufI+FPyIXCcsRbKsueeeFpFOmXFnn3g48vtJ/7LOpzQ433zRQ8a/GiUcle0lH3i4tfL3Mfb+3ZaNqikP6PKLjyPhT8iFwr/AAZ/sW5Y+PhdU6WMv/sfp+8uN980UPGvyZGVLeE95WCPTj09np7PT2ens9PZ6ez09np7PT2ens9PZ6ez09np7PT2ens9PZ6ez09np7PT2ens9PZ6ez09np7PT2ens9PZ6ez09np7PT2ens9PZ6ez09np7PT2ens9PZ6ez09np7PT2ens9PZ6ez09np7PT2ens9PZ6ex2LihrlfGNSSWF1ZnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaZ1pnWmdaWRtuXCn5ELhNc2UVFyza4WPjK95GCxGpUc3losYYhnjffNFDxr+jz2JP3+pyUvIhcby2z+5bftFXnGPKj9xbUHVl7wgorlXC++aKHjX9HnsODydNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnTZ02dNnIyl5ELht7DX+HWsuZ5jK1qQYreoynYN706agsLjffNFDxr6R+yyO+nk18zXzNfM18zXzNfM18zXzNfM18zXzNfM18zXzNfM18zXzNfM18ynfNySP/AH82EYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhGEYRhE/OL8l980UPGvo/8FxLlptj/iLCwylLmgu7KMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMoyjKMon5xfkvvmih419JfS/48fxUWMv+JLuvZNVcHMzmZzM5mczOZnMzmZzM5mczOZnMzmZzM5mczOZnMzmZzM5mczOZnMzmZzM5mczOZnMzmZzM5mczOZnMzmZzM5mczOZnMzmZzM5mczOZnM8lL5i/JffNFDxr6S/l+7H8awluu3/ANL6ll832P8AkpP94tvyXz/ei38a+hq1Y01kd+2zXzKtWVR/xVsUazps18yN/IpVVUXFrJVss/CVGdP5Y4YZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhmGYZhnuzKRb3risNX1Nmuga+Ar+BrqZroGups11M11M11MlfwJz52W/tTX0NzNzke2OHv/ABdjK4ZLSo4z5RdjUXv0oHQgaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemaemdCngrxipGZGOLZQteoslSySWU4tPsyjZlDxr6BlVYmPf+XbrNQW32b9y5f8AyPt3RaTTp4JtRiVZJy7VuUPGvob2GJj/AJdjTzLJ/n7S8pYlntz7YKdSVPaVxOW/t/nspQ5pkI8qx9DUpRqb6CBoIGggaCBoIGggaCBoIGggaCBoIGggaCBoIGggaCBoIGggaCBoIGggaCBoIGggaCBoIGggQpqG32koJjsYMVhA0EDQQNBA0EDQQNBA0EDQQNBA0ECnQUP/AMK//8QAIREBAQAABQUBAQAAAAAAAAAAAQAwMUBgcCBBUICQEKD/2gAIAQMBCT8A4GPxmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZngPLgHLgHtwB386RERERERERERERERERERERERERERERERERERERERERH9KD6A5+gzMzg9t6ERHT31jMzMz0szMzMzMzMzMzMzMzMzMzuNwWfT9jwmW2WcFmZ1OW+zgA+SzwA6TLg0iIiIiIiOeWZnfrMzp2ZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZ8cRERERG6j4JvsDnpWZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmZmelmZmZmZmZmZmZmZmZmZmZmZmcPP6lEb/IiIiIiIiIiIiIiIiIiIiIiIiIiIiIiI+DH/9k=", Bitmap.CompressFormat.JPEG, 0, UtilsFechas.getHoy(Localizacion.getInstance().formatoFechas));
+    public ImagenCommons(Context context) {
+        this(CommonsUtilsImagenes.fromDrawableToBase64Image(context.getDrawable(R.drawable.imagen_demo_comprimida)), Bitmap.CompressFormat.JPEG, 0, UtilsFechas.getHoy(Localizacion.getInstance().formatoFechas));
     }
 
     public ImagenCommons(String contenido, Bitmap.CompressFormat formato, int orientacion) {
